[api] import the requests into the API database
This is a rather huge change and didn't happen overnight. But now the
requests are stored just like the packages and projects within the
same database and the XML is only rendered on request.

The informations of the requests are stored as strings, not as foreign keys
because we delete projects and packages in the database, but (so far) leave
requests for them around.

This is not so bad actually as the database indexes work well enough on such
small strings

+++ src/api/app/controllers/attribute_controller.rb
@@ -166,7 +166,7 @@ class AttributeController < ApplicationController
       render_ok
     elsif request.delete?
       at = ans.attrib_types.where("name = ?", name ).first
-      at.destroy
+      at.destroy if at
       render_ok
     else
       render_error :status => 400, :errorcode => 'illegal_request',
